Added @secure methods, shopcontroller+html page+repository, separated account info from account table, new design login page,

+++ src/main/java/com/app/bank_acquiring/controller/TerminalController.java
@@ -1,19 +1,24 @@
 package com.app.bank_acquiring.controller;
 
-import com.app.bank_acquiring.domain.Account;
+import com.app.bank_acquiring.domain.account.Account;
 import com.app.bank_acquiring.domain.Terminal;
 import com.app.bank_acquiring.repository.AccountRepository;
 import com.app.bank_acquiring.repository.TerminalRepository;
 import com.app.bank_acquiring.service.UposService;
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.util.NoSuchElementException;
 
 @Controller
 public class TerminalController {
@@ -25,29 +30,73 @@ public class TerminalController {
     @Autowired
     private AccountRepository accountRepository;
 
-    @GetMapping("/terminals")
+    /*@GetMapping("/terminals")
     public String list(Model model){
         model.addAttribute("terminals",terminalRepository.findAll());
+        return "terminals";
+    }*/
+
+    @GetMapping("/terminals")
+    public String listAccountTerminals(Model model, @AuthenticationPrincipal UserDetails currentUser) {
+        Account user = accountRepository.findByUsername(currentUser.getUsername());
+        model.addAttribute("account", user);
         return "terminals";
     }
 
     @GetMapping("/terminals/{id}")
-    public String getTerminal(Model model, @PathVariable Long id){
-        model.addAttribute("terminal",terminalRepository.getOne(id));
+    public String getTerminalById(Model model, @PathVariable Long id,
+                              @AuthenticationPrincipal UserDetails currentUser){
+        Account user = accountRepository.findByUsername(currentUser.getUsername());
+        Terminal terminal = terminalRepository.getOne(id);
+        if(user.getTerminals().contains(terminal)){
+            model.addAttribute("term",terminal);
+        } else {
+            throw new RuntimeException("Current account doesn't have this terminal");
+        }
         return "terminal";
     }
 
+   @ModelAttribute
+    private Terminal getTerminal() {
+        return new Terminal();
+    }
+
+    @Secured("ADMIN")
+    @Transactional
     @PostMapping("/terminals")
-    public String add(@ModelAttribute Terminal terminal){
+    public String setAccountTerminals(@Valid @ModelAttribute Terminal terminal, BindingResult bindingResult,
+                                      @AuthenticationPrincipal UserDetails currentUser, Model model) throws IOException {
+        Account acc = accountRepository.findByUsername(currentUser.getUsername());
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("account", acc);
+            return "terminals";
+        }
+        terminal.setAccount(acc);
         terminalRepository.save(terminal);
+        acc.getTerminals().add(terminal);
+        uposService.createUserUpos(acc.getId(), terminal);
         return "redirect:/terminals";
     }
 
+
+    @PostMapping("/terminals/workingterminal")
+    public String setAccountWorkTerminal(@RequestParam Long terminalId,
+                                         @AuthenticationPrincipal UserDetails currentUser) {
+        Account acc = accountRepository.findByUsername(currentUser.getUsername());
+        acc.setWorkTerminalTid(terminalRepository.getOne(terminalId).getTid());
+        accountRepository.save(acc);
+        return "redirect:/terminals";
+    }
+
+    @Secured("ADMIN")
     @PostMapping("/terminals/{id}")
     public String updateTerminal(@RequestParam String ip, @RequestParam String chequeHeader,
                                  @PathVariable Long id, @AuthenticationPrincipal UserDetails currentUser) throws FileNotFoundException {
         Account acc = accountRepository.findByUsername(currentUser.getUsername());
         Terminal terminal = terminalRepository.getOne(id);
+        if(!acc.getTerminals().contains(terminal)){
+            throw new RuntimeException("Current account doesn't have this terminal");
+        }
         if(!ip.isEmpty()){
             terminal.setIp(ip);
         }
@@ -59,12 +108,16 @@ public class TerminalController {
         return "redirect:/terminals/"+id;
     }
 
+    @Secured("ADMIN")
     @DeleteMapping("/terminals/{id}")
     public String deleteTerminal(@PathVariable Long id, @AuthenticationPrincipal UserDetails currentUser) {
         Account acc = accountRepository.findByUsername(currentUser.getUsername());
         Terminal terminal = terminalRepository.getOne(id);
+        if(!acc.getTerminals().contains(terminal)){
+            throw new RuntimeException("Current account doesn't have this terminal");
+        }
         uposService.deleteUserUpos(acc.getId(),terminal.getTid());
         terminalRepository.delete(terminal);
-        return "redirect:/accounts/current";
+        return "redirect:/terminals";
     }
 }
